+++ docs/change_log.rst
@@ -71,11 +71,11 @@ The following are only available within Stem's `git repository
 
  * **Utilities**
 
+  * IPv6 support in :func:`~stem.util.connection.get_connections` when resolving with proc, netstat, lsof, or ss (:trac:`18079`)
   * Added :func:`~stem.util.__init__.datetime_to_unix`
-  * IPv6 support in :func:`~stem.util.connection.get_connections`
   * The 'ss' connection resolver didn't work on Gentoo (:trac:`18079`)
   * Recognize IPv4-mapped IPv6 addresses in our utils (:trac:`18079`)
-  * Added an **is_ipv6** value to Connection instances
+  * Added an **is_ipv6** value to :class:`~stem.util.connection.Connection` instances
   * Allow :func:`stem.util.conf.Config.set` to remove values when provided with a **None** value
 
  * **Interpreter**